+++ vtm-android-example/src/org/oscim/android/test/MarkerOverlayActivity.java
@@ -25,10 +25,9 @@ import org.oscim.backend.canvas.Bitmap;
 import org.oscim.core.GeoPoint;
 import org.oscim.layers.TileGridLayer;
 import org.oscim.layers.marker.ItemizedLayer;
-import org.oscim.layers.marker.ItemizedLayer.OnItemGestureListener;
 import org.oscim.layers.marker.MarkerItem;
-import org.oscim.layers.marker.MarkerItem.HotspotPlace;
 import org.oscim.layers.marker.MarkerSymbol;
+import org.oscim.layers.marker.MarkerSymbol.HotspotPlace;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -37,9 +36,9 @@ import static org.oscim.android.canvas.AndroidGraphics.drawableToBitmap;
 import static org.oscim.tiling.source.bitmap.DefaultSources.STAMEN_TONER;
 
 public class MarkerOverlayActivity extends BitmapTileMapActivity
-        implements OnItemGestureListener<MarkerItem> {
+        implements ItemizedLayer.OnItemGestureListener<MarkerItem> {
 
-    private static final boolean BILLBOARDS = false;
+    private static final boolean BILLBOARDS = true;
     private MarkerSymbol mFocusMarker;
 
     public MarkerOverlayActivity() {
@@ -68,22 +67,21 @@ public class MarkerOverlayActivity extends BitmapTileMapActivity
             mFocusMarker = new MarkerSymbol(drawableToBitmap(d), HotspotPlace.CENTER, false);
 
         ItemizedLayer<MarkerItem> markerLayer =
-                new ItemizedLayer<MarkerItem>(mMap, new ArrayList<MarkerItem>(),
+                new ItemizedLayer<>(mMap, new ArrayList<MarkerItem>(),
                         symbol, this);
 
         mMap.layers().add(markerLayer);
 
-        List<MarkerItem> pts = new ArrayList<MarkerItem>();
+        List<MarkerItem> pts = new ArrayList<>();
 
         for (double lat = -90; lat <= 90; lat += 5) {
             for (double lon = -180; lon <= 180; lon += 5)
-                pts.add(new MarkerItem(lat + "/" + lon, "",
-                        new GeoPoint(lat, lon)));
+                pts.add(new MarkerItem(lat + "/" + lon, "", new GeoPoint(lat, lon)));
         }
 
         markerLayer.addItems(pts);
 
-        mMap.layers().add(new TileGridLayer(mMap));
+        mMap.layers().add(new TileGridLayer(mMap, getResources().getDisplayMetrics().density));
     }
 
     @Override