+++ vtm-android-example/src/org/oscim/android/test/Samples.java
@@ -26,14 +26,10 @@ import android.view.View.OnClickListener;
 import android.widget.Button;
 import android.widget.LinearLayout;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 /**
  * A simple start screen for the sample activities.
  */
 public class Samples extends Activity {
-    public static Logger log = LoggerFactory.getLogger(Samples.class);
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -42,17 +38,21 @@ public class Samples extends Activity {
         setContentView(R.layout.activity_samples);
         LinearLayout linearLayout = (LinearLayout) findViewById(R.id.samples);
         linearLayout.addView(createButton(SimpleMapActivity.class));
+        linearLayout.addView(createButton(MapboxMapActivity.class));
         linearLayout.addView(createButton(BitmapTileMapActivity.class));
         linearLayout.addView(createButton(MapsforgeMapActivity.class));
         linearLayout.addView(createButton(MarkerOverlayActivity.class));
         linearLayout.addView(createButton(PathOverlayActivity.class));
         linearLayout.addView(createButton(LineTexActivity.class));
         linearLayout.addView(createButton(LayerGroupActivity.class));
+        linearLayout.addView(createButton(LocationActivity.class));
         linearLayout.addView(createButton(ThemeStylerActivity.class));
         linearLayout.addView(createButton(S3DBMapActivity.class));
         linearLayout.addView(createButton(JeoIndoorMapActivity.class));
         linearLayout.addView(createButton(OsmJsonMapActivity.class));
         linearLayout.addView(createButton(VectorLayerMapActivity.class));
+        linearLayout.addView(createButton(MultiMapActivity.class));
+        linearLayout.addView(createButton(MapFragmentActivity.class));
     }
 
     private Button createButton(final Class<?> clazz) {