+++ vtm-app/src/org/oscim/app/RouteSearch.java
@@ -28,8 +28,8 @@ import android.widget.TextView;
 import org.oscim.android.canvas.AndroidGraphics;
 import org.oscim.core.GeoPoint;
 import org.oscim.layers.PathLayer;
-import org.oscim.layers.marker.MarkerItem.HotspotPlace;
 import org.oscim.layers.marker.MarkerSymbol;
+import org.oscim.layers.marker.MarkerSymbol.HotspotPlace;
 import org.osmdroid.location.GeocoderNominatim;
 import org.osmdroid.overlays.DefaultInfoWindow;
 import org.osmdroid.overlays.ExtendedMarkerItem;
@@ -427,7 +427,7 @@ public class RouteSearch {
                 time = hour + "h " + minute + "m";
             }
 
-            double dis = ((double) (mStartPoint.distanceTo(mDestinationPoint))) / 1000;
+            double dis = ((double) (mStartPoint.sphericalDistance(mDestinationPoint))) / 1000;
             String distance;
             String shortpath;
             if (dis < 100) {