+++ vtm-app/src/org/oscim/app/TileMap.java
@@ -40,6 +40,7 @@ import org.oscim.app.location.LocationDialog;
 import org.oscim.app.location.LocationHandler;
 import org.oscim.app.preferences.EditPreferences;
 import org.oscim.core.GeoPoint;
+import org.oscim.core.Tile;
 import org.oscim.overlay.DistanceTouchOverlay;
 import org.osmdroid.location.POI;
 import org.osmdroid.overlays.MapEventsReceiver;
@@ -73,6 +74,7 @@ public class TileMap extends MapActivity implements MapEventsReceiver {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
+        Tile.SIZE = Tile.calculateTileSize(getResources().getDisplayMetrics().scaledDensity);
         setContentView(R.layout.activity_tilemap);
         App.view = (MapView) findViewById(R.id.mapView);
         registerMapView(App.view);
@@ -203,7 +205,7 @@ public class TileMap extends MapActivity implements MapEventsReceiver {
                 break;
 
             case R.id.menu_layer_grid:
-                mMapLayers.enableGridOverlay(!mMapLayers.isGridEnabled());
+                mMapLayers.enableGridOverlay(this, !mMapLayers.isGridEnabled());
                 mMap.updateMap(true);
                 break;