+++ vtm-desktop/src/org/oscim/gdx/GdxMapApp.java
@@ -45,28 +45,28 @@ public class GdxMapApp extends GdxMap {
     }
 
     public static void main(String[] args) {
-        Tile.SIZE = 360;
         init();
-        new LwjglApplication(new GdxMapApp(), getConfig());
+        new LwjglApplication(new GdxMapApp(), getConfig(null));
+    }
+
+    public static void run(GdxMap map) {
+        run(map, null, Tile.SIZE);
     }
 
     public static void run(GdxMap map, LwjglApplicationConfiguration config, int tileSize) {
         Tile.SIZE = FastMath.clamp(tileSize, 128, 512);
 
-        new LwjglApplication(map, (config == null ? getConfig() : config));
+        new LwjglApplication(map, (config == null ? getConfig(map.getClass().getSimpleName()) : config));
     }
 
-    public static void run(LwjglApplicationConfiguration config, int tileSize, GdxMap map) {
-        run(map, config, tileSize);
-    }
-
-    static protected LwjglApplicationConfiguration getConfig() {
+    protected static LwjglApplicationConfiguration getConfig(String title) {
+        LwjglApplicationConfiguration.disableAudio = true;
         LwjglApplicationConfiguration cfg = new LwjglApplicationConfiguration();
-        cfg.title = "vtm-gdx";
+        cfg.title = title != null ? title : "vtm-gdx";
         cfg.width = 800;
         cfg.height = 600;
         cfg.stencil = 8;
-        //cfg.samples = 2;
+        cfg.samples = 2;
         cfg.foregroundFPS = 30;
         cfg.backgroundFPS = 10;
         return cfg;
@@ -76,14 +76,7 @@ public class GdxMapApp extends GdxMap {
     public void createLayers() {
         TileSource tileSource = new OSciMap4TileSource();
 
-        // TileSource tileSource = new MapFileTileSource();
-        // tileSource.setOption("file", "/home/jeff/germany.map");
-
         initDefaultLayers(tileSource, false, true, true);
-
-        //mMap.getLayers().add(new BitmapTileLayer(mMap, new ImagicoLandcover(), 20));
-        //mMap.getLayers().add(new BitmapTileLayer(mMap, new OSMTileSource(), 20));
-        //mMap.getLayers().add(new BitmapTileLayer(mMap, new ArcGISWorldShaded(), 20));
 
         mMap.setMapPosition(0, 0, 1 << 2);
     }