+++ vtm-gdx/src/org/oscim/gdx/InputHandler.java
@@ -26,6 +26,7 @@ import org.oscim.layers.TileGridLayer;
 import org.oscim.map.Map;
 import org.oscim.map.ViewController;
 import org.oscim.theme.VtmThemes;
+import org.oscim.utils.Easing;
 
 public class InputHandler implements InputProcessor {
 
@@ -96,9 +97,11 @@ public class InputHandler implements InputProcessor {
                 break;
             case Input.Keys.NUM_1:
                 mMap.animator().animateZoom(500, 0.5, 0, 0);
+                mMap.updateMap(false);
                 break;
             case Input.Keys.NUM_2:
                 mMap.animator().animateZoom(500, 2, 0, 0);
+                mMap.updateMap(false);
                 break;
 
             case Input.Keys.D:
@@ -216,7 +219,6 @@ public class InputHandler implements InputProcessor {
 
     @Override
     public boolean mouseMoved(int screenX, int screenY) {
-
         mPosX = screenX;
         mPosY = screenY;
         return false;
@@ -224,18 +226,10 @@ public class InputHandler implements InputProcessor {
 
     @Override
     public boolean scrolled(int amount) {
-
-        if (amount > 0) {
-
-            mMap.animator().animateZoom(250, 0.75f, 0, 0);
-        } else {
-            float fx = mPosX - mMap.getWidth() / 2;
-            float fy = mPosY - mMap.getHeight() / 2;
-
-            mMap.animator().animateZoom(250, 1.333f, fx, fy);
-        }
+        float fx = mPosX - mMap.getWidth() / 2;
+        float fy = mPosY - mMap.getHeight() / 2;
+        mMap.animator().animateZoom(250, amount > 0 ? 0.75f : 1.333f, fx, fy, Easing.Type.LINEAR);
         mMap.updateMap(false);
-
         return true;
     }
 }