+++ vtm/src/org/oscim/backend/CanvasAdapter.java
@@ -44,9 +44,19 @@ public abstract class CanvasAdapter {
     static CanvasAdapter g;
 
     /**
+     * Default dpi.
+     */
+    public static final float DEFAULT_DPI = 160;
+
+    /**
      * The dpi.
      */
-    public static float dpi = 240;
+    public static float dpi = DEFAULT_DPI;
+
+    /**
+     * The scale.
+     */
+    public static float scale = 1;
 
     /**
      * The text scale.