+++ vtm/src/org/oscim/map/Viewport.java
@@ -1,6 +1,7 @@
 /*
  * Copyright 2012 Hannes Janetzek
  * Copyright 2016 devemux86
+ * Copyright 2016 Erik Duisters
  *
  * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
  *
@@ -25,24 +26,20 @@ import org.oscim.core.Point;
 import org.oscim.core.Tile;
 import org.oscim.renderer.GLMatrix;
 import org.oscim.utils.FastMath;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * The Viewport class contains a MapPosition and the projection matrices.
  * It provides functions to modify the MapPosition and translate between
  * map and screen coordinates.
- * <p/>
+ * <p>
  * Public methods are thread safe.
  */
 public class Viewport {
 
-    static final Logger log = LoggerFactory.getLogger(Viewport.class);
-
-    private final static int MAX_ZOOMLEVEL = 20;
-    private final static int MIN_ZOOMLEVEL = 2;
-    private final static float MIN_TILT = 0;
-    private final static float MAX_TILT = 65;
+    public final static int MAX_ZOOMLEVEL = 20;
+    public final static int MIN_ZOOMLEVEL = 2;
+    public final static float MIN_TILT = 0;
+    public final static float MAX_TILT = 65;
 
     protected double mMaxScale = (1 << MAX_ZOOMLEVEL);
     protected double mMinScale = (1 << MIN_ZOOMLEVEL);
@@ -331,14 +328,33 @@ public class Viewport {
     }
 
     /**
-     * Get the screen pixel for a GeoPoint
+     * Get the screen pixel for a GeoPoint (relative to center)
      *
      * @param geoPoint the GeoPoint
      * @param out      Point projected to screen pixel relative to center
      */
     public void toScreenPoint(GeoPoint geoPoint, Point out) {
+        toScreenPoint(geoPoint, true, out);
+    }
+
+    /**
+     * Get the screen pixel for a GeoPoint
+     *
+     * @param geoPoint the GeoPoint
+     * @param out      Point projected to screen pixel
+     */
+    public void toScreenPoint(GeoPoint geoPoint, boolean relativeToCenter, Point out) {
         MercatorProjection.project(geoPoint, out);
-        toScreenPoint(out.x, out.y, out);
+        toScreenPoint(out.x, out.y, relativeToCenter, out);
+    }
+
+    /**
+     * Get the screen pixel for map coordinates (relative to center)
+     *
+     * @param out Point projected to screen coordinate relative to center
+     */
+    public void toScreenPoint(double x, double y, Point out) {
+        toScreenPoint(x, y, true, out);
     }
 
     /**
@@ -346,7 +362,7 @@ public class Viewport {
      *
      * @param out Point projected to screen coordinate
      */
-    public void toScreenPoint(double x, double y, Point out) {
+    public void toScreenPoint(double x, double y, boolean relativeToCenter, Point out) {
 
         double cs = mPos.scale * Tile.SIZE;
         double cx = mPos.x * cs;
@@ -362,9 +378,20 @@ public class Viewport {
 
         out.x = (mv[0] * (mWidth / 2));
         out.y = -(mv[1] * (mHeight / 2));
+
+        if (!relativeToCenter) {
+            out.x += mWidth / 2;
+            out.y += mHeight / 2;
+        }
+    }
+
+    boolean sizeChanged(Viewport viewport) {
+        return mHeight != viewport.mHeight || mWidth != viewport.mWidth;
     }
 
     protected boolean copy(Viewport viewport) {
+        boolean sizeChanged = sizeChanged(viewport);
+
         mHeight = viewport.mHeight;
         mWidth = viewport.mWidth;
         mProjMatrix.copy(viewport.mProjMatrix);
@@ -375,7 +402,7 @@ public class Viewport {
         mRotationMatrix.copy(viewport.mRotationMatrix);
         mViewMatrix.copy(viewport.mViewMatrix);
         mViewProjMatrix.copy(viewport.mViewProjMatrix);
-        return viewport.getMapPosition(mPos);
+        return viewport.getMapPosition(mPos) || sizeChanged;
     }
 
     public double getMaxScale() {