+++ docs/Changelog.md
@@ -1,20 +1,56 @@
 # Changelog
 
-## New for 0.6.0
+## New since 0.6.0
 
-**Revive of VTM vector map library.**
+- Mapsforge multiple map files [#208](https://github.com/mapsforge/vtm/issues/208)
+- Polygon label position enhancements [#80](https://github.com/mapsforge/vtm/issues/80)
+- vtm-web modules update [#51](https://github.com/mapsforge/vtm/issues/51)
+- Mapbox vector tiles [#57](https://github.com/mapsforge/vtm/issues/57)
+- PathLayer (vtm) fix disappearing segments [#108](https://github.com/mapsforge/vtm/issues/108)
+- House numbers (nodes) fix visibility [#168](https://github.com/mapsforge/vtm/issues/168)
+- Use baseline 160dpi in scaling [#236](https://github.com/mapsforge/vtm/issues/236)
+- MapFileTileSource zoom level API enhancements [#219](https://github.com/mapsforge/vtm/issues/219)
+- Animator enhancements with easing functions [#246](https://github.com/mapsforge/vtm/issues/246)
+- Tile grid layer scaling [#238](https://github.com/mapsforge/vtm/issues/238)
+- Internal render themes various improvements [#41](https://github.com/mapsforge/vtm/issues/41)
+- Many other minor improvements and bug fixes
+- [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aissue+is%3Aclosed+milestone%3A0.7.0)
+
+## Version 0.6.X
+
+### Version 0.6.0 (2016-10-28) -- VTM revive
+
+- Same as 0.6.0-rc3
+
+#### Version 0.6.0-rc3 (2016-10-22)
+
+- Minor improvements and bug fixes
+
+#### Version 0.6.0-rc2 (2016-10-16)
+
+- Location layer [#171](https://github.com/mapsforge/vtm/issues/171)
+- Minor improvements and bug fixes
+
+#### Version 0.6.0-rc1 (2016-10-08)
 
 - Render theme SVG resources [#60](https://github.com/mapsforge/vtm/issues/60)
 - Mapsforge multilingual maps [#34](https://github.com/mapsforge/vtm/issues/34)
-- vtm-ios update module [#29](https://github.com/mapsforge/vtm/issues/29)
+- Render theme styles [#93](https://github.com/mapsforge/vtm/issues/93)
+- vtm-ios module update [#29](https://github.com/mapsforge/vtm/issues/29)
 - Native libraries for all platforms [#14](https://github.com/mapsforge/vtm/issues/14)
-- LWJGL desktop libGDX backend [#129](https://github.com/mapsforge/vtm/issues/129)
 - Line stipple and texture rendering [#105](https://github.com/mapsforge/vtm/issues/105)
 - Layer groups [#99](https://github.com/mapsforge/vtm/issues/99) [#103](https://github.com/mapsforge/vtm/issues/103)
-- Map scale bar multi-platform [#84](https://github.com/mapsforge/vtm/issues/84)
+- Location renderer [#171](https://github.com/mapsforge/vtm/issues/171)
+- Map scale bar [#84](https://github.com/mapsforge/vtm/issues/84)
+- Tile size based on scale factor [#183](https://github.com/mapsforge/vtm/issues/183)
+- libGDX layer gestures [#151](https://github.com/mapsforge/vtm/issues/151)
 - Render theme area tessellation option [#37](https://github.com/mapsforge/vtm/issues/37)
 - Graphics API platform enhancements [#92](https://github.com/mapsforge/vtm/issues/92)
+- GeoPoint & BoundingBox improvements [#201](https://github.com/mapsforge/vtm/issues/201) [#200](https://github.com/mapsforge/vtm/issues/200)
 - vtm-jts module [#53](https://github.com/mapsforge/vtm/issues/53)
-- Internal render themes various enhancements
+- vtm-http module [#140](https://github.com/mapsforge/vtm/issues/140)
+- Internal render themes various improvements [#41](https://github.com/mapsforge/vtm/issues/41)
+- LWJGL desktop libGDX backend [#129](https://github.com/mapsforge/vtm/issues/129)
+- SNAPSHOT builds publish to Sonatype OSSRH [#165](https://github.com/mapsforge/vtm/issues/165)
 - Many other minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aissue+is%3Aclosed+milestone%3A0.6.0)